Added feedback for when you try to enter an inconsistent value.
The offending cell and any conflicts will all flash red until you
release the button.
Numbers are now placed in cells on UIEventTouchDown instead of
TouchUpInside, which is going to be really annoying if we want to make
the buttons actually click.
In order for the GridView to be able to unflash-red the conflicting
cells, CellView now has an additional setTarget method for a "flash
timeout" -- GridView sets itself as the target of this and then
whenever that happens, it goes through all of its cells and unflashes
them.